een verwijderknop toegevoegd en uc9 afgerond

+++ Grocery.Core/Services/GroceryListItemsService.cs
@@ -34,13 +34,12 @@ namespace Grocery.Core.Services
             return _groceriesRepository.Add(item);
         }
 
-        // UC09: Implementatie van Delete methode
         public GroceryListItem? Delete(GroceryListItem item)
         {
+            if (item == null) return null;
             return _groceriesRepository.Delete(item);
         }
 
-        // UC09: Implementatie van Delete methode met ID
         public GroceryListItem? Delete(int id)
         {
             var item = _groceriesRepository.Get(id);
@@ -51,21 +50,20 @@ namespace Grocery.Core.Services
             return null;
         }
 
-        // UC09: Implementatie van Get methode
         public GroceryListItem? Get(int id)
         {
             var item = _groceriesRepository.Get(id);
             if (item != null)
             {
-                // Fill product information
                 item.Product = _productRepository.Get(item.ProductId) ?? new(0, "", 0);
             }
             return item;
         }
 
-        // UC09: Implementatie van Update methode
+        
         public GroceryListItem? Update(GroceryListItem item)
         {
+            if (item == null) return null;
             return _groceriesRepository.Update(item);
         }
